# api/openapi.yaml
openapi: '3.0.3'
info:
  title: 'Motocorp Dealership Management System API'
  description: |
    A comprehensive cloud-based dealership management system for motorcycle dealerships.
    This API provides multi-tenant architecture with complete CRUD operations for:
    - Inventory Management (motorcycles, parts, accessories)
    - Customer Relationship Management (CRM)
    - Sales Pipeline Management
    - Service Appointment Management
    - User Management with role-based access control
    - Real-time notifications and activity feeds
    
    ## Authentication
    This API uses JWT Bearer token authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your_jwt_token>`
    
    ## Multi-Tenancy
    The API supports multi-tenant architecture where each organization has its own isolated data.
    The tenant context is automatically resolved from the authenticated user's JWT token.
    
    ## Rate Limiting
    API requests are rate-limited to ensure fair usage and system stability.
    
    ## Error Handling
    Standard HTTP status codes are used. Error responses include detailed error messages
    and validation errors where applicable.
  version: '1.0.0'
  contact:
    name: 'Motocorp DMS Support'
    email: 'support@motocorpdms.com'
    url: 'https://motocorpdms.com/support'
  license:
    name: 'Proprietary'
servers:
  - url: 'https://api.motocorpdms.com/v1'
    description: 'Production server'
...
    description: 'Staging server'
...
security:
  - BearerAuth: []
paths:
...
  /dealerships:
    get:
      tags:
        - Dealerships
      summary: List dealerships
      description: Retrieves a paginated list of dealerships within the organization
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search
          in: query
          description: Search dealerships by name
          schema:
            type: string
        - name: isActive
          in: query
          description: Filter by active status
          schema:
            type: boolean
        - name: timezone
          in: query
          description: Filter by timezone
          schema:
            type: string
        - name: sortBy
          in: query
          description: Field to sort by
          schema:
            type: string
            default: 'createdAt'
        - name: sortOrder
          in: query
          description: Sort direction
          schema:
            type: string
            enum: [asc, desc]
            default: 'desc'
      responses:
        '200':
          description: Dealerships retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDealerships'

    post:
      tags:
        - Dealerships
      summary: Create a new dealership
      description: Creates a new dealership within the organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDealershipDto'
      responses:
        '201':
          description: Dealership created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealership'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /dealerships/{id}:
    get:
      tags:
        - Dealerships
      summary: Get dealership by ID
      description: Retrieves a specific dealership by its ID
      parameters:
        - $ref: '#/components/parameters/DealershipIdParam'
      responses:
        '200':
          description: Dealership retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealership'
        '404':
          description: Dealership not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      tags:
        - Dealerships
      summary: Update dealership
      description: Updates an existing dealership's information
      parameters:
        - $ref: '#/components/parameters/DealershipIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDealershipDto'
      responses:
        '200':
          description: Dealership updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dealership'
        '404':
          description: Dealership not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Dealerships
      summary: Delete dealership
      description: Soft deletes a dealership (sets isActive to false)
      parameters:
        - $ref: '#/components/parameters/DealershipIdParam'
      responses:
        '204':
          description: Dealership deleted successfully
        '404':
          description: Dealership not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /dealerships/{id}/stats:
    get:
      tags:
        - Dealerships
      summary: Get dealership statistics
      description: Retrieves aggregated statistics for a specific dealership
      parameters:
        - $ref: '#/components/parameters/DealershipIdParam'
      responses:
        '200':
          description: Dealership statistics retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DealershipStats'
        '404':
          description: Dealership not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # User Management
...
  /customers:
    get:
      tags:
        - Customers
      summary: List customers
      description: Retrieves a paginated list of customers with optional filtering
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: search
          in: query
          description: Search customers by name, email, or phone
          schema:
            type: string
        - name: dealershipId
          in: query
          description: Filter customers by dealership
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Customers retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCustomers'

    post:
      tags:
        - Customers
      summary: Create a new customer
      description: Creates a new customer record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerDto'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
...
  /customers/{id}:
    get:
      tags:
        - Customers
      summary: Get customer by ID
      description: Retrieves a specific customer with their complete profile
      parameters:
        - $ref: '#/components/parameters/CustomerIdParam'
      responses:
        '200':
          description: Customer retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerDetail'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      tags:
        - Customers
      summary: Update customer
      description: Updates an existing customer's information
      parameters:
        - $ref: '#/components/parameters/CustomerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCustomerDto'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags:
        - Customers
      summary: Delete customer
      description: Permanently deletes a customer record
      parameters:
        - $ref: '#/components/parameters/CustomerIdParam'
      responses:
        '204':
          description: Customer deleted successfully
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /customers/{id}/transfer:
    patch:
      tags:
        - Customers
      summary: Transfer customer to another dealership
      description: Transfers a customer from one dealership to another
      parameters:
        - $ref: '#/components/parameters/CustomerIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferCustomerDto'
      responses:
        '200':
          description: Customer transferred successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Inventory Management
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT Bearer token authentication. Format: Bearer <your_jwt_token>'
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination (starts from 1)
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    CustomerIdParam:
      name: id
      in: path
      required: true
      description: Customer ID
      schema:
        type: string
        format: uuid
    DealershipIdParam:
      name: id
      in: path
      required: true
      description: Dealership ID
      schema:
        type: string
        format: uuid
  schemas:
...
    CustomerDetail:
      allOf:
        - $ref: '#/components/schemas/Customer'
        - type: object
          properties:
            recentSales:
              type: array
              items:
                $ref: '#/components/schemas/Sale'
              description: Recent sales for this customer
            upcomingAppointments:
              type: array
              items:
                $ref: '#/components/schemas/ServiceAppointment'
              description: Upcoming service appointments
    CreateCustomerDto:
      type: object
      required:
        - firstName
        - lastName
      properties:
        firstName:
          type: string
          minLength: 2
          description: Customer's first name
          example: 'John'
        lastName:
          type: string
          minLength: 2
          description: Customer's last name
          example: 'Smith'
        email:
          type: string
          format: email
          nullable: true
          description: Customer's email address
          example: 'john.smith@example.com'
        phone:
          type: string
          nullable: true
          description: Customer's phone number
          example: '+1-555-123-4567'
        primaryDealershipId:
          type: string
          format: uuid
          nullable: true
          description: Primary dealership ID
          example: '123e4567-e89b-12d3-a456-426614174002'
        address:
          type: object
          nullable: true
          properties:
            street:
              type: string
              example: '123 Main St'
            city:
              type: string
              example: 'Springfield'
            state:
              type: string
              example: 'IL'
            zipCode:
              type: string
              example: '62701'
            country:
              type: string
              example: 'USA'
        preferences:
          type: object
          nullable: true
          properties:
            preferredContactMethod:
              type: string
              enum: [email, phone, sms]
              example: 'email'
            marketingOptIn:
              type: boolean
              example: true
            serviceReminders:
              type: boolean
              example: true
            newsletter:
              type: boolean
              example: false
            preferredLanguage:
              type: string
              example: 'en'
...
    TransferCustomerDto:
      type: object
      required:
        - newDealershipId
      properties:
        newDealershipId:
          type: string
          format: uuid
          description: ID of the dealership to transfer customer to
          example: '123e4567-e89b-12d3-a456-426614174003'
        reason:
          type: string
          nullable: true
          description: Reason for transfer
          example: 'Customer relocated'
    PaginatedCustomers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
...
    Dealership:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique dealership identifier
          example: '123e4567-e89b-12d3-a456-426614174000'
        organizationId:
          type: string
          format: uuid
          description: Organization ID
          example: '123e4567-e89b-12d3-a456-426614174001'
        name:
          type: string
          description: Dealership name
          example: 'Downtown Motorcycles'
        location:
          type: object
          nullable: true
          properties:
            address:
              type: string
              example: '123 Main Street'
            city:
              type: string
              example: 'New York'
            state:
              type: string
              example: 'NY'
            zipCode:
              type: string
              example: '10001'
            country:
              type: string
              example: 'USA'
            coordinates:
              type: object
              properties:
                lat:
                  type: number
                  example: 40.7128
                lng:
                  type: number
                  example: -74.0060
          description: Dealership location details
        timezone:
          type: string
          description: Dealership timezone
          example: 'America/New_York'
          default: 'UTC'
        businessHours:
          type: object
          nullable: true
          properties:
            monday:
              type: object
              properties:
                open:
                  type: string
                  example: '09:00'
                close:
                  type: string
                  example: '18:00'
                closed:
                  type: boolean
                  example: false
            tuesday:
              type: object
              properties:
                open:
                  type: string
                  example: '09:00'
                close:
                  type: string
                  example: '18:00'
                closed:
                  type: boolean
                  example: false
            wednesday:
              type: object
              properties:
                open:
                  type: string
                  example: '09:00'
                close:
                  type: string
                  example: '18:00'
                closed:
                  type: boolean
                  example: false
            thursday:
              type: object
              properties:
                open:
                  type: string
                  example: '09:00'
                close:
                  type: string
                  example: '18:00'
                closed:
                  type: boolean
                  example: false
            friday:
              type: object
              properties:
                open:
                  type: string
                  example: '09:00'
                close:
                  type: string
                  example: '18:00'
                closed:
                  type: boolean
                  example: false
            saturday:
              type: object
              properties:
                open:
                  type: string
                  example: '10:00'
                close:
                  type: string
                  example: '17:00'
                closed:
                  type: boolean
                  example: false
            sunday:
              type: object
              properties:
                closed:
                  type: boolean
                  example: true
          description: Business hours for each day of the week
        isActive:
          type: boolean
          description: Whether the dealership is active
          example: true
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
          example: '2024-01-01T00:00:00Z'
        updatedAt:
          type: string
          format: date-time
          description: Last update timestamp
          example: '2024-01-15T10:30:00Z'
        _count:
          type: object
          description: Related entity counts
          properties:
            users:
              type: integer
              example: 12
            inventory:
              type: integer
              example: 156
            customers:
              type: integer
              example: 245
            sales:
              type: integer
              example: 23
            serviceAppointments:
              type: integer
              example: 18
    CreateDealershipDto:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          description: Dealership name
          example: 'Westside Motors'
        location:
          type: object
          nullable: true
          properties:
            address:
              type: string
              example: '456 Oak Avenue'
            city:
              type: string
              example: 'Los Angeles'
            state:
              type: string
              example: 'CA'
            zipCode:
              type: string
              example: '90210'
            country:
              type: string
              example: 'USA'
            coordinates:
              type: object
              properties:
                lat:
                  type: number
                  example: 34.0522
                lng:
                  type: number
                  example: -118.2437
          description: Dealership location details
        timezone:
          type: string
          description: Dealership timezone
          example: 'America/Los_Angeles'
          default: 'UTC'
        businessHours:
          type: object
          nullable: true
          properties:
            monday:
              type: object
              properties:
                open:
                  type: string
                  example: '08:30'
                close:
                  type: string
                  example: '19:00'
                closed:
                  type: boolean
                  example: false
            tuesday:
              type: object
              properties:
                open:
                  type: string
                  example: '08:30'
                close:
                  type: string
                  example: '19:00'
                closed:
                  type: boolean
                  example: false
            wednesday:
              type: object
              properties:
                open:
                  type: string
                  example: '08:30'
                close:
                  type: string
                  example: '19:00'
                closed:
                  type: boolean
                  example: false
            thursday:
              type: object
              properties:
                open:
                  type: string
                  example: '08:30'
                close:
                  type: string
                  example: '19:00'
                closed:
                  type: boolean
                  example: false
            friday:
              type: object
              properties:
                open:
                  type: string
                  example: '08:30'
                close:
                  type: string
                  example: '19:00'
                closed:
                  type: boolean
                  example: false
            saturday:
              type: object
              properties:
                open:
                  type: string
                  example: '09:00'
                close:
                  type: string
                  example: '18:00'
                closed:
                  type: boolean
                  example: false
            sunday:
              type: object
              properties:
                closed:
                  type: boolean
                  example: true
          description: Business hours for each day of the week
        isActive:
          type: boolean
          description: Whether the dealership is active
          example: true
          default: true
    UpdateDealershipDto:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          description: Dealership name
          example: 'Westside Motors & Parts'
        location:
          type: object
          nullable: true
          properties:
            address:
              type: string
              example: '789 Pine Street'
            city:
              type: string
              example: 'Los Angeles'
            state:
              type: string
              example: 'CA'
            zipCode:
              type: string
              example: '90211'
            country:
              type: string
              example: 'USA'
            coordinates:
              type: object
              properties:
                lat:
                  type: number
                  example: 34.0522
                lng:
                  type: number
                  example: -118.2437
          description: Dealership location details
        timezone:
          type: string
          description: Dealership timezone
          example: 'America/Los_Angeles'
        businessHours:
          type: object
          nullable: true
          properties:
            monday:
              type: object
              properties:
                open:
                  type: string
                  example: '08:00'
                close:
                  type: string
                  example: '20:00'
                closed:
                  type: boolean
                  example: false
            tuesday:
              type: object
              properties:
                open:
                  type: string
                  example: '08:00'
                close:
                  type: string
                  example: '20:00'
                closed:
                  type: boolean
                  example: false
            wednesday:
              type: object
              properties:
                open:
                  type: string
                  example: '08:00'
                close:
                  type: string
                  example: '20:00'
                closed:
                  type: boolean
                  example: false
            thursday:
              type: object
              properties:
                open:
                  type: string
                  example: '08:00'
                close:
                  type: string
                  example: '20:00'
                closed:
                  type: boolean
                  example: false
            friday:
              type: object
              properties:
                open:
                  type: string
                  example: '08:00'
                close:
                  type: string
                  example: '20:00'
                closed:
                  type: boolean
                  example: false
            saturday:
              type: object
              properties:
                open:
                  type: string
                  example: '09:00'
                close:
                  type: string
                  example: '18:00'
                closed:
                  type: boolean
                  example: false
            sunday:
              type: object
              properties:
                open:
                  type: string
                  example: '11:00'
                close:
                  type: string
                  example: '16:00'
                closed:
                  type: boolean
                  example: false
          description: Business hours for each day of the week
        isActive:
          type: boolean
          description: Whether the dealership is active
          example: true
    PaginatedDealerships:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Dealership'
        meta:
          type: object
          properties:
            total:
              type: integer
              description: Total number of dealerships
              example: 5
            page:
              type: integer
              description: Current page number
              example: 1
            limit:
              type: integer
              description: Items per page
              example: 10
            totalPages:
              type: integer
              description: Total number of pages
              example: 1
    DealershipStats:
      type: object
      properties:
        users:
          type: object
          properties:
            total:
              type: integer
              description: Total users in dealership
              example: 12
            active:
              type: integer
              description: Active users in dealership
              example: 11
        inventory:
          type: object
          properties:
            total:
              type: integer
              description: Total inventory items
              example: 156
            lowStock:
              type: integer
              description: Items below reorder point
              example: 8
        customers:
          type: integer
          description: Total customers
          example: 245
        sales:
          type: object
          properties:
            active:
              type: integer
              description: Active sales opportunities
              example: 23
        serviceAppointments:
          type: object
          properties:
            scheduled:
              type: integer
              description: Scheduled appointments
              example: 18

    # Organization Schema
...
    PaginationMeta:
      type: object
      properties:
        currentPage:
          type: integer
          description: Current page number
          example: 1
        totalPages:
          type: integer
          description: Total number of pages
          example: 5
        totalItems:
          type: integer
          description: Total number of items
          example: 95
        itemsPerPage:
          type: integer
          description: Number of items per page
          example: 20
        hasNextPage:
          type: boolean
          description: Whether there is a next page
          example: true
        hasPreviousPage:
          type: boolean
          description: Whether there is a previous page
          example: false
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Human readable error message
          example: 'Resource not found'
        error:
          type: string
          description: Error type
          example: 'NotFound'
        statusCode:
          type: integer
          description: HTTP status code
          example: 404
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: '2024-01-15T10:30:00Z'
        path:
          type: string
          description: Request path that caused the error
          example: '/api/v1/customers/invalid-id'
    ValidationError:
      type: object
      required:
        - message
        - errors
      properties:
        message:
          type: string
          description: General validation error message
          example: 'Validation failed'
        error:
          type: string
          description: Error type
          example: 'Bad Request'
        statusCode:
          type: integer
          description: HTTP status code
          example: 400
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
          example: '2024-01-15T10:30:00Z'
        path:
          type: string
          description: Request path that caused the error
          example: '/api/v1/customers'
        errors:
          type: array
          description: Detailed validation errors
          items:
            type: object
            properties:
              field:
                type: string
                description: Field that failed validation
                example: 'email'
              message:
                type: string
                description: Validation error message
                example: 'must be a valid email address'
              value:
                description: Invalid value that was provided
                example: 'invalid-email'
...